Refactor `new_test_image()` and `delete_test_image()` as a context manager.

+++ icekit/utils/testing.py
@@ -1,54 +1,51 @@
-# USEFUL FUNCTIONS DESIGNED FOR TESTS ##############################################################
+import contextlib
 import glob
 import os
 import uuid
 
 from django.conf import settings
-from django.core.files.base import ContentFile
 from PIL import Image, ImageDraw
-from six import BytesIO
 
 
-def new_test_image():
+@contextlib.contextmanager
+def get_test_image(storage):
     """
-    Creates an automatically generated test image.
+    Context manager that creates an image with the given storage class, returns
+    a storage name, and cleans up (including thumbnails) when done.
 
-    In your testing `tearDown` method make sure to delete the test
-    image with the helper function `delete_test_image`.
+    Example:
 
-    The recommended way of using this helper function is as follows:
+        with get_test_image(obj.image_field.storage) as name:
+            obj.image_field = name
 
-        object_1.image_property.save(*new_test_image())
-
-    :return: Image name and image content file.
     """
-    image_name = 'test-{}.png'.format(uuid.uuid4())
-    image = Image.new('RGBA', size=(50, 50), color=(256, 0, 0))
-    ImageDraw.Draw(image)
-    byte_io = BytesIO()
-    image.save(byte_io, 'png')
-    byte_io.seek(0)
+    # Generate unique name.
+    storage_name = 'test-{}.png'.format(uuid.uuid4())
 
-    return image_name, ContentFile(byte_io.read(), image_name)
+    # Create image.
+    image_file = Image.new('RGBA', size=(50, 50), color=(255, 0, 0))
+    ImageDraw.Draw(image_file)
+    storage_file = storage.open(storage_name, 'w')
+    image_file.save(storage_file, 'png')
 
+    # Yield storage name.
+    yield storage_name
 
-def delete_test_image(image_field):
-    """
-    Deletes test image generated as well as thumbnails if created.
+    # Delete thumbnails, if they exist.
+    THUMBNAIL_ROOT = os.path.join(
+        settings.MEDIA_ROOT,
+        settings.THUMBNAIL_BASEDIR,
+    )
+    if hasattr(settings, 'THUMBNAIL_BASEDIR'):
+        thumbs = glob.glob(os.path.join(THUMBNAIL_ROOT, storage_name + '*'))
+        for filename in thumbs:
+            os.unlink(filename)
 
-    The recommended way of using this helper function is as follows:
+    # Delete empty thumbnail directories.
+    try:
+        os.removedirs(THUMBNAIL_ROOT)
+    except OSError:
+        pass
 
-        delete_test_image(object_1.image_property)
-
-    :param image_field: The image field on an object.
-    :return: None.
-    """
-    # ensure all thumbs are deleted
-    for filename in glob.glob(
-        os.path.join(
-            settings.BASE_DIR, 'public', 'media', 'thumbs', image_field.name.split('/')[-1]
-        ) + '*'
-    ):
-        os.unlink(filename)
-    # delete the saved file
-    image_field.delete()
+    # Delete image.
+    storage.delete(storage_name)
